[ADD] prototypes

+++ lib/my/my_compute_power_rec.c
@@ -5,7 +5,7 @@
 ** my compute power rec
 */
 
-int my_compute_power_rec(int nb , int p)
+int my_compute_power_rec(long int nb , int p)
 {
     if (p == 0) {
         return 1;
